end  graphQl week1 Apollo and codegen

+++ frontend/src/components/admin/AdminTagRow.tsx
@@ -6,24 +6,24 @@ import { PencilSquareIcon } from "@heroicons/react/24/outline";
 import { CheckIcon } from "@heroicons/react/24/outline";
 import { XCircleIcon } from "@heroicons/react/24/outline";
 import axios from "axios";
+import { useUpdateTagMutation } from "@/graphql/generated/schema";
 
 interface AdminTagRowProps {
   tag: Tag;
   handleDeleteTag: (id: number) => Promise<void>;
 }
-export default function AdminTagRow({
-  tag: { id, name },
-  handleDeleteTag,
-}: AdminTagRowProps) {
+export default function AdminTagRow({ tag: { id, name }, handleDeleteTag }: AdminTagRowProps) {
   const [isEditing, setIsEditing] = useState(false);
 
   const [displayedName, setDisplayedName] = useState(name);
 
+  const [updateTag] = useUpdateTagMutation();
+
   const handleSave = async () => {
     try {
       if (displayedName) {
-        await axios.patch(`http://localhost:4000/tags/${id}`, {
-          name: displayedName,
+        await updateTag({
+          variables: { data: { name: displayedName }, tagId: id },
         });
         setIsEditing(false);
       }
@@ -37,16 +37,7 @@ export default function AdminTagRow({
     <tr>
       <td>{id}</td>
       <td>
-        {isEditing ? (
-          <input
-            type="text"
-            className="input"
-            value={displayedName}
-            onChange={(e) => setDisplayedName(e.target.value)}
-          />
-        ) : (
-          displayedName
-        )}
+        {isEditing ? <input type="text" className="input" value={displayedName} onChange={(e) => setDisplayedName(e.target.value)} /> : displayedName}
       </td>
       <td>
         {isEditing ? (
@@ -70,8 +61,7 @@ export default function AdminTagRow({
             </button>
             <button
               onClick={() => {
-                if (confirm("voulez vous vraiement supprimer le tag ?"))
-                  handleDeleteTag(id);
+                if (confirm("voulez vous vraiement supprimer le tag ?")) handleDeleteTag(id);
               }}
             >
               <TrashIcon width={24} height={24} className="ml-2" />
